+++ src/lib/renderers/customFunctions.js
@@ -1,4 +1,7 @@
 export const customFunctions = {
     "Math": Math,
-    "Number": Number
+    "Number": Number,
+    "log": function () {
+        console.log('you are logging from an AG grid function');
+    }
 }